+++ ipc.c
@@ -187,8 +187,8 @@ initialize_web_extensions_cb(WebKitWebContext *context, gpointer UNUSED(data))
     g_free(dirs[0]);
 }
 
-static void
-remove_socket_file(void)
+void
+ipc_remove_socket_file(void)
 {
     g_mutex_lock(&socket_path_lock);
     g_unlink(socket_path);
@@ -204,8 +204,7 @@ ipc_init(void)
     g_thread_new("accept_thread", web_extension_connect_thread, NULL);
     g_signal_connect(web_context_get(), "initialize-web-extensions",
             G_CALLBACK (initialize_web_extensions_cb), NULL);
-    /* Remove socket file at exit */
-    atexit(remove_socket_file);
+    atexit(ipc_remove_socket_file);
 }
 
 // vim: ft=c:et:sw=4:ts=8:sts=4:tw=80