+++ luah.c
@@ -19,6 +19,7 @@
  *
  */
 
+#include "ipc.h"
 #include "luah.h"
 #include "log.h"
 #include "common/luah.h"
@@ -262,6 +263,7 @@ luaH_parserc(const gchar *confpath, gboolean run)
         setenv("LUAKIT_QUEUED_EMISSIONS_FILE", log_dump_file, TRUE);
         g_free(log_dump_file);
     }
+    ipc_remove_socket_file();
     execvp(escaped_execpath, (gchar**)argv->pdata);
 
 bailout: