+++ CHANGELOG.md
@@ -32,6 +32,7 @@
  - Unique instance support has been moved to `unique_instance.lua`.
  - The `image` widget now uses `luakit.resource_path` to locate local files.
  - The log viewer now shows errors logged by a user-defined rc.lua failing to load.
+ - Luakit will now remove its IPC socket file before restarting.
 
 ### Removed